+++ changelog.md
@@ -59,6 +59,8 @@
 * GMC feed: Availability value "available for order" deprecated
 * Mobile: Shopping cart warnings weren't displayed to customers
 * Tax provider and payment method were automatically activated when there were no active provider\method
+* #784 Biz-Importer: Name of delivery time must not be imported empty
+* #776 Preview: Manufacturer and Product in Multi Store
 
 
 ## SmartStore.NET 2.2.2